move GraphDetails to pages folder
